requirements.txt
streamlit
langchain
openai
elevenlabs
SpeechRecognition
pyaudio
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gspread
oauth2client
python-dotenv
langchain-community
